+++ lib/api/src/import_object.rs
@@ -6,8 +6,7 @@ use std::collections::VecDeque;
 use std::collections::{hash_map::Entry, HashMap};
 use std::fmt;
 use std::sync::{Arc, Mutex};
-use wasmer_engine::NamedResolver;
-use wasmer_vm::Export;
+use wasmer_engine::{Export, NamedResolver};
 
 /// The `LikeNamespace` trait represents objects that act as a namespace for imports.
 /// For example, an `Instance` or `Namespace` could be
@@ -265,7 +264,6 @@ mod test {
     use crate::{Global, Store, Val};
     use wasmer_engine::ChainableNamedResolver;
     use wasmer_types::Type;
-    use wasmer_vm::Export;
 
     #[test]
     fn chaining_works() {
@@ -320,7 +318,7 @@ mod test {
         let happy_dog_entry = resolver.resolve_by_name("dog", "happy").unwrap();
 
         assert!(if let Export::Global(happy_dog_global) = happy_dog_entry {
-            happy_dog_global.from.ty().ty == Type::I64
+            happy_dog_global.vm_global.from.ty().ty == Type::I64
         } else {
             false
         });
@@ -346,7 +344,7 @@ mod test {
         let happy_dog_entry = resolver.resolve_by_name("dog", "happy").unwrap();
 
         assert!(if let Export::Global(happy_dog_global) = happy_dog_entry {
-            happy_dog_global.from.ty().ty == Type::I32
+            happy_dog_global.vm_global.from.ty().ty == Type::I32
         } else {
             false
         });
@@ -366,7 +364,7 @@ mod test {
         let happy_dog_entry = imports1.resolve_by_name("dog", "happy").unwrap();
 
         assert!(if let Export::Global(happy_dog_global) = happy_dog_entry {
-            happy_dog_global.from.ty().ty == Type::I32
+            happy_dog_global.vm_global.from.ty().ty == Type::I32
         } else {
             false
         });