+++ lib/api/src/instance.rs
@@ -2,9 +2,10 @@ use crate::exports::Exports;
 use crate::externals::Extern;
 use crate::module::Module;
 use crate::store::Store;
-use crate::InstantiationError;
+use crate::{HostEnvInitError, LinkError, RuntimeError};
 use std::fmt;
 use std::sync::{Arc, Mutex};
+use thiserror::Error;
 use wasmer_engine::Resolver;
 use wasmer_vm::{InstanceHandle, VMContext};
 
@@ -35,6 +36,44 @@ mod send_test {
     #[test]
     fn instance_is_send() {
         assert!(is_send::<Instance>());
+    }
+}
+
+/// An error while instantiating a module.
+///
+/// This is not a common WebAssembly error, however
+/// we need to differentiate from a `LinkError` (an error
+/// that happens while linking, on instantiation), a
+/// Trap that occurs when calling the WebAssembly module
+/// start function, and an error when initializing the user's
+/// host environments.
+#[derive(Error, Debug)]
+pub enum InstantiationError {
+    /// A linking ocurred during instantiation.
+    #[error(transparent)]
+    Link(LinkError),
+
+    /// A runtime error occured while invoking the start function
+    #[error(transparent)]
+    Start(RuntimeError),
+
+    /// Error occurred when initializing the host environment.
+    #[error(transparent)]
+    HostEnvInitialization(HostEnvInitError),
+}
+
+impl From<wasmer_engine::InstantiationError> for InstantiationError {
+    fn from(other: wasmer_engine::InstantiationError) -> Self {
+        match other {
+            wasmer_engine::InstantiationError::Link(e) => Self::Link(e),
+            wasmer_engine::InstantiationError::Start(e) => Self::Start(e),
+        }
+    }
+}
+
+impl From<HostEnvInitError> for InstantiationError {
+    fn from(other: HostEnvInitError) -> Self {
+        Self::HostEnvInitialization(other)
     }
 }
 
@@ -80,16 +119,34 @@ impl Instance {
             .map(|export| {
                 let name = export.name().to_string();
                 let export = handle.lookup(&name).expect("export");
-                let extern_ = Extern::from_export(store, export);
+                let extern_ = Extern::from_export(store, export.into());
                 (name, extern_)
             })
             .collect::<Exports>();
 
-        Ok(Self {
+        let instance = Self {
             handle: Arc::new(Mutex::new(handle)),
             module: module.clone(),
             exports,
-        })
+        };
+
+        // # Safety
+        // `initialize_host_envs` should be called after instantiation but before
+        // returning an `Instance` to the user. We set up the host environments
+        // via `WasmerEnv::init_with_instance`.
+        //
+        // This usage is correct because we pass a valid pointer to `instance` and the
+        // correct error type returned by `WasmerEnv::init_with_instance` as a generic
+        // parameter.
+        unsafe {
+            instance
+                .handle
+                .lock()
+                .unwrap()
+                .initialize_host_envs::<HostEnvInitError>(&instance as *const _ as *const _)?;
+        }
+
+        Ok(instance)
     }
 
     /// Gets the [`Module`] associated with this instance.