+++ lib/api/tests/externals.rs
@@ -208,8 +208,9 @@ fn function_new() -> Result<()> {
 #[test]
 fn function_new_env() -> Result<()> {
     let store = Store::default();
-    #[derive(Clone)]
+    #[derive(Clone, WasmerEnv)]
     struct MyEnv {};
+
     let my_env = MyEnv {};
     let function = Function::new_native_with_env(&store, my_env.clone(), |_env: &MyEnv| {});
     assert_eq!(function.ty().clone(), FunctionType::new(vec![], vec![]));
@@ -270,7 +271,7 @@ fn function_new_dynamic() -> Result<()> {
 #[test]
 fn function_new_dynamic_env() -> Result<()> {
     let store = Store::default();
-    #[derive(Clone)]
+    #[derive(Clone, WasmerEnv)]
     struct MyEnv {};
     let my_env = MyEnv {};
 
@@ -373,6 +374,36 @@ fn function_outlives_instance() -> Result<()> {
     };
 
     assert_eq!(f.call(4, 5)?, 9);
+
+    Ok(())
+}
+
+#[test]
+fn manually_generate_wasmer_env() -> Result<()> {
+    let store = Store::default();
+    #[derive(WasmerEnv)]
+    struct MyEnv {
+        val: u32,
+        memory: LazyInit<Memory>,
+    }
+
+    fn host_function(env: &mut MyEnv, arg1: u32, arg2: u32) -> u32 {
+        env.val + arg1 + arg2
+    }
+
+    let mut env = MyEnv {
+        val: 5,
+        memory: LazyInit::new(),
+    };
+
+    let result = host_function(&mut env, 7, 9);
+    assert_eq!(result, 21);
+
+    let memory = Memory::new(&store, MemoryType::new(0, None, false))?;
+    env.memory.initialize(memory);
+
+    let result = host_function(&mut env, 1, 2);
+    assert_eq!(result, 8);
 
     Ok(())
 }