+++ lib/vm/src/export.rs
@@ -10,23 +10,23 @@ use wasmer_types::{FunctionType, MemoryType, TableType};
 
 /// The value of an export passed from one instance to another.
 #[derive(Debug, Clone)]
-pub enum Export {
+pub enum VMExport {
     /// A function export value.
-    Function(ExportFunction),
+    Function(VMExportFunction),
 
     /// A table export value.
-    Table(ExportTable),
+    Table(VMExportTable),
 
     /// A memory export value.
-    Memory(ExportMemory),
+    Memory(VMExportMemory),
 
     /// A global export value.
-    Global(ExportGlobal),
+    Global(VMExportGlobal),
 }
 
 /// A function export value.
 #[derive(Debug, Clone, PartialEq)]
-pub struct ExportFunction {
+pub struct VMExportFunction {
     /// The address of the native-code function.
     pub address: *const VMFunctionBody,
     /// Pointer to the containing `VMContext`.
@@ -43,20 +43,20 @@ pub struct ExportFunction {
 /// # Safety
 /// There is no non-threadsafe logic directly in this type. Calling the function
 /// may not be threadsafe.
-unsafe impl Send for ExportFunction {}
+unsafe impl Send for VMExportFunction {}
 /// # Safety
-/// The members of an ExportFunction are immutable after construction.
-unsafe impl Sync for ExportFunction {}
+/// The members of an VMExportFunction are immutable after construction.
+unsafe impl Sync for VMExportFunction {}
 
-impl From<ExportFunction> for Export {
-    fn from(func: ExportFunction) -> Self {
+impl From<VMExportFunction> for VMExport {
+    fn from(func: VMExportFunction) -> Self {
         Self::Function(func)
     }
 }
 
 /// A table export value.
 #[derive(Debug, Clone)]
-pub struct ExportTable {
+pub struct VMExportTable {
     /// Pointer to the containing `Table`.
     pub from: Arc<dyn Table>,
 }
@@ -65,14 +65,14 @@ pub struct ExportTable {
 /// This is correct because there is no non-threadsafe logic directly in this type;
 /// correct use of the raw table from multiple threads via `definition` requires `unsafe`
 /// and is the responsibilty of the user of this type.
-unsafe impl Send for ExportTable {}
+unsafe impl Send for VMExportTable {}
 /// # Safety
 /// This is correct because the values directly in `definition` should be considered immutable
 /// and the type is both `Send` and `Clone` (thus marking it `Sync` adds no new behavior, it
 /// only makes this type easier to use)
-unsafe impl Sync for ExportTable {}
+unsafe impl Sync for VMExportTable {}
 
-impl ExportTable {
+impl VMExportTable {
     /// Get the table type for this exported table
     pub fn ty(&self) -> &TableType {
         self.from.ty()
@@ -83,21 +83,21 @@ impl ExportTable {
         self.from.style()
     }
 
-    /// Returns whether or not the two `ExportTable`s refer to the same Memory.
+    /// Returns whether or not the two `VMExportTable`s refer to the same Memory.
     pub fn same(&self, other: &Self) -> bool {
         Arc::ptr_eq(&self.from, &other.from)
     }
 }
 
-impl From<ExportTable> for Export {
-    fn from(table: ExportTable) -> Self {
+impl From<VMExportTable> for VMExport {
+    fn from(table: VMExportTable) -> Self {
         Self::Table(table)
     }
 }
 
 /// A memory export value.
 #[derive(Debug, Clone)]
-pub struct ExportMemory {
+pub struct VMExportMemory {
     /// Pointer to the containing `Memory`.
     pub from: Arc<dyn Memory>,
 }
@@ -106,14 +106,14 @@ pub struct ExportMemory {
 /// This is correct because there is no non-threadsafe logic directly in this type;
 /// correct use of the raw memory from multiple threads via `definition` requires `unsafe`
 /// and is the responsibilty of the user of this type.
-unsafe impl Send for ExportMemory {}
+unsafe impl Send for VMExportMemory {}
 /// # Safety
 /// This is correct because the values directly in `definition` should be considered immutable
 /// and the type is both `Send` and `Clone` (thus marking it `Sync` adds no new behavior, it
 /// only makes this type easier to use)
-unsafe impl Sync for ExportMemory {}
+unsafe impl Sync for VMExportMemory {}
 
-impl ExportMemory {
+impl VMExportMemory {
     /// Get the type for this exported memory
     pub fn ty(&self) -> &MemoryType {
         self.from.ty()
@@ -124,21 +124,21 @@ impl ExportMemory {
         self.from.style()
     }
 
-    /// Returns whether or not the two `ExportMemory`s refer to the same Memory.
+    /// Returns whether or not the two `VMExportMemory`s refer to the same Memory.
     pub fn same(&self, other: &Self) -> bool {
         Arc::ptr_eq(&self.from, &other.from)
     }
 }
 
-impl From<ExportMemory> for Export {
-    fn from(memory: ExportMemory) -> Self {
+impl From<VMExportMemory> for VMExport {
+    fn from(memory: VMExportMemory) -> Self {
         Self::Memory(memory)
     }
 }
 
 /// A global export value.
 #[derive(Debug, Clone)]
-pub struct ExportGlobal {
+pub struct VMExportGlobal {
     /// The global declaration, used for compatibility checking.
     pub from: Arc<Global>,
 }
@@ -147,22 +147,22 @@ pub struct ExportGlobal {
 /// This is correct because there is no non-threadsafe logic directly in this type;
 /// correct use of the raw global from multiple threads via `definition` requires `unsafe`
 /// and is the responsibilty of the user of this type.
-unsafe impl Send for ExportGlobal {}
+unsafe impl Send for VMExportGlobal {}
 /// # Safety
 /// This is correct because the values directly in `definition` should be considered immutable
 /// from the perspective of users of this type and the type is both `Send` and `Clone` (thus
 /// marking it `Sync` adds no new behavior, it only makes this type easier to use)
-unsafe impl Sync for ExportGlobal {}
+unsafe impl Sync for VMExportGlobal {}
 
-impl ExportGlobal {
-    /// Returns whether or not the two `ExportGlobal`s refer to the same Global.
+impl VMExportGlobal {
+    /// Returns whether or not the two `VMExportGlobal`s refer to the same Global.
     pub fn same(&self, other: &Self) -> bool {
         Arc::ptr_eq(&self.from, &other.from)
     }
 }
 
-impl From<ExportGlobal> for Export {
-    fn from(global: ExportGlobal) -> Self {
+impl From<VMExportGlobal> for VMExport {
+    fn from(global: VMExportGlobal) -> Self {
         Self::Global(global)
     }
 }