+++ tests/compilers/native_functions.rs
@@ -78,6 +78,90 @@ fn native_function_works_for_wasm() -> Result<()> {
     Ok(())
 }
 
+#[test]
+#[should_panic(
+    expected = "Closures (functions with captured environments) are currently unsupported with native functions. See: https://github.com/wasmerio/wasmer/issues/1840"
+)]
+fn native_host_function_closure_panics() {
+    let store = get_store(false);
+    let state = 3;
+    Function::new_native(&store, move |_: i32| {
+        println!("{}", state);
+    });
+}
+
+#[test]
+#[should_panic(
+    expected = "Closures (functions with captured environments) are currently unsupported with native functions. See: https://github.com/wasmerio/wasmer/issues/1840"
+)]
+fn native_with_env_host_function_closure_panics() {
+    let store = get_store(false);
+    let state = 3;
+    let env = 4;
+    Function::new_native_with_env(&store, env, move |_env: &i32, _: i32| {
+        println!("{}", state);
+    });
+}
+
+#[test]
+fn non_native_functions_and_closures_with_no_env_work() -> Result<()> {
+    let store = get_store(false);
+    let wat = r#"(module
+        (func $multiply1 (import "env" "multiply1") (param i32 i32) (result i32))
+        (func $multiply2 (import "env" "multiply2") (param i32 i32) (result i32))
+        (func $multiply3 (import "env" "multiply3") (param i32 i32) (result i32))
+        (func $multiply4 (import "env" "multiply4") (param i32 i32) (result i32))
+
+        (func (export "test") (param i32 i32 i32 i32 i32) (result i32)
+           (call $multiply4
+             (call $multiply3
+               (call $multiply2
+                  (call $multiply1
+                    (local.get 0)
+                    (local.get 1))
+                  (local.get 2))
+               (local.get 3))
+              (local.get 4)))
+)"#;
+    let module = Module::new(&store, wat).unwrap();
+
+    let ty = FunctionType::new(vec![Type::I32, Type::I32], vec![Type::I32]);
+    let env = 10;
+    let captured_by_closure = 20;
+    let import_object = imports! {
+        "env" => {
+            "multiply1" => Function::new(&store, &ty, move |args| {
+                if let (Value::I32(v1), Value::I32(v2)) = (&args[0], &args[1]) {
+                    Ok(vec![Value::I32(v1 * v2 * captured_by_closure)])
+                } else {
+                    panic!("Invalid arguments");
+                }
+            }),
+            "multiply2" => Function::new_with_env(&store, &ty, env, move |&env, args| {
+                if let (Value::I32(v1), Value::I32(v2)) = (&args[0], &args[1]) {
+                    Ok(vec![Value::I32(v1 * v2 * captured_by_closure * env)])
+                } else {
+                    panic!("Invalid arguments");
+                }
+            }),
+            "multiply3" => Function::new_native(&store, |arg1: i32, arg2: i32| -> i32
+                                                {arg1 * arg2 }),
+            "multiply4" => Function::new_native_with_env(&store, env, |&env: &i32, arg1: i32, arg2: i32| -> i32
+                                                         {arg1 * arg2 * env }),
+        },
+    };
+
+    let instance = Instance::new(&module, &import_object)?;
+
+    let test: NativeFunc<(i32, i32, i32, i32, i32), i32> =
+        instance.exports.get_native_function("test")?;
+
+    let result = test.call(2, 3, 4, 5, 6)?;
+    let manually_computed_result = 6 * (5 * (4 * (3 * 2 * 20) * 10 * 20)) * 10;
+    assert_eq!(result, manually_computed_result);
+    Ok(())
+}
+
 // The native ABI for functions fails when defining a function natively in
 // macos (Darwin) with the Apple Silicon ARM chip
 // TODO: Cranelift should have a good ABI for the ABI
@@ -244,8 +328,15 @@ fn static_host_function_with_env() -> anyhow::Result<()> {
         Ok((d * 4.0, c * 3.0, b * 2, a * 1))
     }
 
-    #[derive(Clone)]
+    #[derive(WasmerEnv, Clone)]
     struct Env(Rc<RefCell<i32>>);
+
+    impl std::ops::Deref for Env {
+        type Target = Rc<RefCell<i32>>;
+        fn deref(&self) -> &Self::Target {
+            &self.0
+        }
+    }
 
     // Native static host function that returns a tuple.
     {
@@ -311,8 +402,15 @@ fn dynamic_host_function_without_env() -> anyhow::Result<()> {
 fn dynamic_host_function_with_env() -> anyhow::Result<()> {
     let store = get_store(false);
 
-    #[derive(Clone)]
+    #[derive(WasmerEnv, Clone)]
     struct Env(Rc<RefCell<i32>>);
+
+    impl std::ops::Deref for Env {
+        type Target = Rc<RefCell<i32>>;
+        fn deref(&self) -> &Self::Target {
+            &self.0
+        }
+    }
 
     let env = Env(Rc::new(RefCell::new(100)));
     let f = Function::new_with_env(